% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connect}
\alias{connect}
\alias{no.results}
\title{Database utilities}
\usage{
connect(dbuser = "notaR", dbpass = "notaR", dbname = "notaR")

no.results(object)
}
\arguments{
\item{dbuser}{Database user}

\item{dbpass}{Database password}

\item{dbname}{MySQL database name}

\item{object}{The result from a database query}
}
\description{
Function used to connect with the database with default or specified credentials. Usually, this
will be used internally by the other functions

Accessory function to test if a database query returned no results
}
\examples{
con = connect()
}
