% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\docType{data}
\name{PATH}
\alias{PATH}
\title{Global configuration}
\format{An object of class \code{character} of length 1.}
\usage{
PATH
}
\description{
The PATH global variable must be defined when notaR is used by a web interface, in order
to correctly set up the data files. For local use, leave it defined as ".".
}
\keyword{datasets}
