% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{PATH}
\alias{PATH}
\title{Variavel global usada pela interface Web}
\format{An object of class \code{character} of length 1.}
\usage{
PATH
}
\description{
A variavel PATH deve ser definida na interface web para carregar os arquivos de dados.
Para uso local, mantenha o PATH definido como "."
}
\keyword{datasets}
