% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suporte.R
\name{equivalent}
\alias{equivalent}
\title{Suporte à correção}
\usage{
equivalent(a, b)
}
\description{
A função \code{equivalent} verifica se dois objetos são equivalentes. 
Deve ser usada no lugar de "identical" para testes que admitem 
pequenas inconsistências numéricas
}
\examples{
identical (1, 1 / 3 * (1 + 1 + 1))
equivalent (1, 1 / 3 * (1 + 1 + 1))
}
