% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{equivalent}
\alias{equivalent}
\title{Support to creating exercises}
\usage{
equivalent(a, b)
}
\arguments{
\item{a, b}{Objects to be compared}
}
\description{
The \code{equivalent} function verifies whether two objects are
similar up to rounding errors. It should be used instead of \code{identical}
for tests that admit a small numerical inconsistency.
}
\examples{
identical (1, 1 / 3 * (1 + 1 + 1))
equivalent (1, 1 / 3 * (1 + 1 + 1))
}
